requirements.txt
streamlit
openai
finnhub-python